chore: Introducing FormValidatableValueItem
